Change the group id org.ballerinalang -> io.ballerina.stdlib in native code
