+++ ballerina/build.gradle
@@ -62,7 +62,6 @@ ballerina {
     packageOrganization = packageOrg
     module = packageName
     langVersion = ballerinaLangVersion
-    testCoverageParam = '--code-coverage --coverage-format=xml --includes=org.ballerinalang.postgresql.*:ballerinax.postgresql.*'
 }
 
 dependencies {
