+++ native/src/main/java/io/ballerina/stdlib/postgresql/utils/Utils.java
@@ -15,13 +15,13 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package org.ballerinalang.postgresql.utils;
+package io.ballerina.stdlib.postgresql.utils;
 
 import io.ballerina.runtime.api.creators.ValueCreator;
 import io.ballerina.runtime.api.values.BDecimal;
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BString;
-import org.ballerinalang.postgresql.Constants;
+import io.ballerina.stdlib.postgresql.Constants;
 
 /**
  * This class includes utility functions.
